edge/requirements.txt
python-dotenv==1.0.1
pyzmq==25.1.2
